+++ .github/CHANGELOG.md
@@ -1,4 +1,17 @@
-# Release 0.4.0dev0
+# Release 0.5.0-dev
+
+* Added controlled rotation gates to PennyLane operations and `default.qubit` plugin.
+  [#251](https://github.com/XanaduAI/pennylane/pull/251)
+
+### Contributors
+
+This release contains contributions from (in alphabetical order):
+
+Aroosa Ijaz
+
+
+
+# Release 0.4.0
 
 ### New features since last release